Switched to using WORD/UWORD instead of L and UL.

+++ arith/number_of_partitions_vec.c
@@ -36,20 +36,20 @@ arith_number_of_partitions_vec(fmpz * res, slong len)
 
     tmp = _fmpz_vec_init(len);
 
-    tmp[0] = 1L;
+    tmp[0] = WORD(1);
 
     for (n = k = 1; n + 4*k + 2 < len; k += 2)
     {
-        tmp[n] = -1L;
-        tmp[n + k] = -1L;
-        tmp[n + 3*k + 1] = 1L;
-        tmp[n + 4*k + 2] = 1L;
+        tmp[n] = WORD(-1);
+        tmp[n + k] = WORD(-1);
+        tmp[n + 3*k + 1] = WORD(1);
+        tmp[n + 4*k + 2] = WORD(1);
         n += 6*k + 5;
     }
 
-    if (n < len) tmp[n] = -1L;
-    if (n + k < len) tmp[n + k] = -1L;
-    if (n + 3*k + 1 < len) tmp[n + 3*k + 1] = 1L;
+    if (n < len) tmp[n] = WORD(-1);
+    if (n + k < len) tmp[n + k] = WORD(-1);
+    if (n + 3*k + 1 < len) tmp[n + 3*k + 1] = WORD(1);
 
     _fmpz_poly_inv_series(res, tmp, len);
