+++ arith/test/t-pi_chudnovsky.c
@@ -41,8 +41,8 @@ int main(void)
     {
         mpfr_t x, y;
 
-        mpfr_init2(x, 1L << k);
-        mpfr_init2(y, 1L << k);
+        mpfr_init2(x, WORD(1) << k);
+        mpfr_init2(y, WORD(1) << k);
 
         mpfr_const_pi(x, MPFR_RNDN);
         mpfr_pi_chudnovsky(y, MPFR_RNDN);
@@ -50,7 +50,7 @@ int main(void)
         if (!mpfr_equal_p(x, y))
         {
             flint_printf("FAIL:\n");
-            flint_printf("Wrong value at prec = %wd\n", 1L << k);
+            flint_printf("Wrong value at prec = %wd\n", WORD(1) << k);
             abort();
         }
