+++ fmpq_poly/compose.c
@@ -34,11 +34,11 @@ void
 _fmpq_poly_compose(fmpz * res, fmpz_t den, const fmpz * poly1, const fmpz_t den1, 
                    slong len1, const fmpz * poly2, const fmpz_t den2, slong len2)
 {
-    if (*den2 == 1L)
+    if (*den2 == WORD(1))
     {
         _fmpz_poly_compose(res, poly1, len1, poly2, len2);
         fmpz_set(den, den1);
-        _fmpq_poly_canonicalise(res, den, (len1 - 1L) * (len2 - 1L) + 1L);
+        _fmpq_poly_canonicalise(res, den, (len1 - WORD(1)) * (len2 - WORD(1)) + WORD(1));
     }
     else
     {
@@ -49,7 +49,7 @@ _fmpq_poly_compose(fmpz * res, fmpz_t den, const fmpz * poly1, const fmpz_t den1
         
         _fmpq_poly_rescale(v, den, poly1, den1, len1, one, den2);
         _fmpz_poly_compose(res, v, len1, poly2, len2);
-        _fmpq_poly_canonicalise(res, den, (len1 - 1L) * (len2 - 1L) + 1L);
+        _fmpq_poly_canonicalise(res, den, (len1 - WORD(1)) * (len2 - WORD(1)) + WORD(1));
         
         fmpz_clear(one);
         _fmpz_vec_clear(v, len1);
@@ -64,12 +64,12 @@ fmpq_poly_compose(fmpq_poly_t res,
     const slong len2 = poly2->length;
     slong lenr;
     
-    if (len1 == 0L)
+    if (len1 == WORD(0))
     {
         fmpq_poly_zero(res);
         return;
     }
-    if (len1 == 1L || len2 == 0L)
+    if (len1 == WORD(1) || len2 == WORD(0))
     {
         fmpq_poly_fit_length(res, 1);
         fmpz_set(res->coeffs, poly1->coeffs);
@@ -78,7 +78,7 @@ fmpq_poly_compose(fmpq_poly_t res,
             fmpz_t d;
             fmpz_init(d);
             fmpz_gcd(d, res->coeffs, res->den);
-            if (*d != 1L)
+            if (*d != WORD(1))
             {
                 fmpz_divexact(res->coeffs, res->coeffs, d);
                 fmpz_divexact(res->den, res->den, d);
@@ -90,7 +90,7 @@ fmpq_poly_compose(fmpq_poly_t res,
         return;
     }
     
-    lenr = (len1 - 1L) * (len2 - 1L) + 1L;
+    lenr = (len1 - WORD(1)) * (len2 - WORD(1)) + WORD(1);
     
     if ((res != poly1) && (res != poly2))
     {
