+++ fmpq_poly/compose_series_brent_kung.c
@@ -103,7 +103,7 @@ _fmpq_poly_compose_series_brent_kung(fmpz * res, fmpz_t den, const fmpz * poly1,
     }
 
     /* Set rows of A to powers of poly2 */
-    fmpq_set_si(fmpq_mat_entry(A, 0, 0), 1L, 1L);
+    fmpq_set_si(fmpq_mat_entry(A, 0, 0), WORD(1), WORD(1));
 
     for (i = 0; i < len2; i++)
     {
