+++ fmpq_poly/resultant.c
@@ -156,7 +156,7 @@ void fmpq_poly_resultant(fmpq_t r, const fmpq_poly_t f, const fmpq_poly_t g)
                                  g->coeffs, g->den, len2, 
                                  f->coeffs, f->den, len1);
 
-            if (((len1 | len2) & 1L) == 0L)
+            if (((len1 | len2) & WORD(1)) == WORD(0))
                 fmpq_neg(r, r);
         }
     }
