+++ fmpq_poly/revert_series_newton.c
@@ -66,7 +66,7 @@ _fmpq_poly_revert_series_newton(fmpz * Qinv, fmpz_t den,
         fmpz_init(Vden);
 
         k = n;
-        for (i = 1; (1L << i) < k; i++);
+        for (i = 1; (WORD(1) << i) < k; i++);
         a = (slong *) flint_malloc(i * sizeof(slong));
         a[i = 0] = k;
         while (k >= FLINT_REVERSE_NEWTON_CUTOFF)
