+++ fmpq_poly/scalar_div_fmpz.c
@@ -33,7 +33,7 @@
 void _fmpq_poly_scalar_div_fmpz(fmpz * rpoly, fmpz_t rden, const fmpz * poly, 
                                 const fmpz_t den, slong len, const fmpz_t c)
 {
-    if (*c == 1L)
+    if (*c == WORD(1))
     {
         if (rpoly != poly)
         {
@@ -41,7 +41,7 @@ void _fmpq_poly_scalar_div_fmpz(fmpz * rpoly, fmpz_t rden, const fmpz * poly,
             fmpz_set(rden, den);
         }
     }
-    else if (*c == -1L)
+    else if (*c == WORD(-1))
     {
         _fmpz_vec_neg(rpoly, poly, len);
         fmpz_set(rden, den);
@@ -65,7 +65,7 @@ void _fmpq_poly_scalar_div_fmpz(fmpz * rpoly, fmpz_t rden, const fmpz * poly,
 
 void fmpq_poly_scalar_div_fmpz(fmpq_poly_t rop, const fmpq_poly_t op, const fmpz_t c)
 {
-    if (*c == 0L)
+    if (*c == WORD(0))
     {
         flint_printf("Exception (fmpq_poly_scalar_div_fmpz). Division by zero.\n");
         abort();
