+++ fmpq_poly/scalar_mul_fmpz.c
@@ -44,9 +44,9 @@ void _fmpq_poly_scalar_mul_fmpz(fmpz * rpoly, fmpz_t rden,
 
     fmpz_init(gcd);
     fmpz_one(gcd);
-    if (*c != 1L)
+    if (*c != WORD(1))
         fmpz_gcd(gcd, c, den);
-    if (*gcd == 1L)
+    if (*gcd == WORD(1))
     {
         _fmpz_vec_scalar_mul_fmpz(rpoly, poly, len, c);
         fmpz_set(rden, den);
