+++ fmpq_poly/scalar_mul_si.c
@@ -46,7 +46,7 @@ void _fmpq_poly_scalar_mul_si(fmpz * rpoly, fmpz_t rden,
     fmpz_init(gcd);
     fmpz_set_si(gcd, c);
     fmpz_gcd(gcd, gcd, den);
-    if (*gcd == 1L)
+    if (*gcd == WORD(1))
     {
         _fmpz_vec_scalar_mul_si(rpoly, poly, len, c);
         fmpz_set(rden, den);
