+++ fmpz/add_ui.c
@@ -35,7 +35,7 @@ void fmpz_add_ui(fmpz_t f, const fmpz_t g, ulong x)
 	if (!COEFF_IS_MPZ(c))  /* g is small */
 	{
         mp_limb_t sum[2];
-		if (c >= 0L)  /* both operands non-negative */
+		if (c >= WORD(0))  /* both operands non-negative */
 		{
 			add_ssaaaa(sum[1], sum[0], 0, c, 0, x);
             fmpz_set_uiui(f, sum[1], sum[0]);
