+++ fmpz/fdiv_r.c
@@ -50,18 +50,18 @@ fmpz_fdiv_r(fmpz_t f, const fmpz_t g, const fmpz_t h)
             fmpz q = c1 / c2;   /* compute C quotient */
             fmpz r = c1 - c2 * q;   /* compute remainder */
 
-            if ((c2 > 0L && r < 0L) || (c2 < 0L && r > 0L))
+            if ((c2 > WORD(0) && r < WORD(0)) || (c2 < WORD(0) && r > WORD(0)))
                 r += c2;
 
             fmpz_set_si(f, r);
         }
         else                    /* h is large and g is small */
         {
-            if (c1 == 0L)
+            if (c1 == WORD(0))
             {
                 fmpz_set_si(f, c1);
             }
-            else if ((c1 < 0L && fmpz_sgn(h) < 0) || (c1 > 0L && fmpz_sgn(h) > 0))  /* signs are the same */
+            else if ((c1 < WORD(0) && fmpz_sgn(h) < 0) || (c1 > WORD(0) && fmpz_sgn(h) > 0))  /* signs are the same */
             {
                 fmpz_set_si(f, c1);
             }
