+++ fmpz/mul_tdiv_q_2exp.c
@@ -45,7 +45,7 @@ fmpz_mul_tdiv_q_2exp(fmpz_t f, const fmpz_t g, const fmpz_t h, ulong exp)
 
     c2 = *h;                    /* save h in case it is aliased with f */
 
-    if (c2 == 0L)               /* special case, h = 0  */
+    if (c2 == WORD(0))               /* special case, h = 0  */
     {
         fmpz_zero(f);
         return;
